requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
SQLAlchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
asyncpg==0.29.0

# Redis for caching
redis==5.0.1
hiredis==2.2.3

# Pydantic for data validation
pydantic[email]==2.5.0
pydantic-settings==2.0.3

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Security
python-keycloak==3.7.0
bcrypt==4.1.2

# Monitoring and observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-prometheus==1.12.0rc1
prometheus-client==0.19.0

# Logging
structlog==23.2.0
rich==13.7.0

# Email
emails==0.6
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
faker==20.1.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
celery==5.3.4
flower==2.0.1

# Image processing (for hotel/room images)
Pillow==10.1.0

# Geolocation
geopy==2.4.0

# Date/time utilities
python-dateutil==2.8.2
pytz==2023.3

# Excel export (for reports)
openpyxl==3.1.2

# Development server reload
watchfiles==0.21.0
